openmodelica 1.9.3 (new formula)
New formula for OpenModelica which a open source Modelica based modeling and
simulation language.
This is a new formula for something which has been available for a while
on macports. Unfortunately the stable version does not include the library
files in the binary thus they cannot be installed on stable, but it does
work on HEAD.